+++ lib/sh/makepath.c
@@ -1,6 +1,6 @@
 /* makepath.c - glue PATH and DIR together into a full pathname. */
 
-/* Copyright (C) 1987-2020,2022 Free Software Foundation, Inc.
+/* Copyright (C) 1987-2020,2022-2023 Free Software Foundation, Inc.
 
    This file is part of GNU Bash, the Bourne Again SHell.
 
@@ -69,7 +69,7 @@ static char *nullpath = "";
 char *
 sh_makepath (const char *path, const char *dir, int flags)
 {
-  int dirlen, pathlen;
+  size_t dirlen, pathlen;
   char *ret, *xpath, *xdir, *r, *s;
 
   if (path == 0 || *path == '\0')
